// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/auditPlanAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AuditCategoryAC } from './auditCategoryAC';
import { AuditPlanDocumentAC } from './auditPlanDocumentAC';
import { AuditPlanSectionType } from './auditPlanSectionType';
import { AuditPlanStatus } from './auditPlanStatus';
import { AuditTypeAC } from './auditTypeAC';
import { BaseModelAC } from './baseModelAC';
import { PlanProcessMappingAC } from './planProcessMappingAC';
import { ProcessAC } from './processAC';

export interface AuditPlanAC extends BaseModelAC { 
    title?: string;
    selectedTypeId?: string;
    auditTypeName?: string;
    selectCategoryId?: string;
    auditCategoryName?: string;
    entityId: string;
    endDateTime?: Date;
    startDateTime?: Date;
    status: AuditPlanStatus;
    statusString?: string;
    version: number;
    overviewBackground?: string;
    totalBudgetedHours?: number;
    auditTypeSelectionDisaplyList?: Array<AuditTypeAC>;
    auditCategorySelectionDisaplyList?: Array<AuditCategoryAC>;
    planProcessList?: Array<PlanProcessMappingAC>;
    parentProcessList?: Array<ProcessAC>;
    planDocumentList?: Array<AuditPlanDocumentAC>;
    sectionType: AuditPlanSectionType;
    financialYear: number;
    parentPlanId?: string;
    endDateTimeToString?: string;
    startDateTimeToString?: string;
}