// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/planProcessMappingAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BaseModelAC } from './baseModelAC';
import { PlanProcessStatus } from './planProcessStatus';

export interface PlanProcessMappingAC extends BaseModelAC { 
    planId: string;
    processId?: string;
    parentProcessId?: string;
    startDateTime: Date;
    auditPlan?: string;
    parentProcessName?: string;
    processName?: string;
    statusString?: string;
    startDateTimeForDisplay?: string;
    endDateTime: Date;
    endDateTimeForDisplay?: string;
    status: PlanProcessStatus;
    workProgramId?: string;
}