// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/observationType.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type ObservationType = 0 | 1 | 2 | 3;

export const ObservationType = {
    NUMBER_0: 0 as ObservationType,
    NUMBER_1: 1 as ObservationType,
    NUMBER_2: 2 as ObservationType,
    NUMBER_3: 3 as ObservationType
};